refactor: move schema classes to their own namespace
