+++ src/Properties/ModelPropertyExtension.php
@@ -7,6 +7,9 @@ namespace NunoMaduro\Larastan\Properties;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Str;
 use Iterator;
+use NunoMaduro\Larastan\Schema\SchemaAggregator;
+use NunoMaduro\Larastan\Schema\SchemaColumn;
+use NunoMaduro\Larastan\Schema\SchemaTable;
 use PHPStan\Parser\CachedParser;
 use PHPStan\PhpDoc\TypeStringResolver;
 use PHPStan\Reflection\ClassReflection;
